+++ packages/abis/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @enzymefinance/abis
 
+## 1.2.6
+
+### Patch Changes
+
+- [#681](https://github.com/enzymefinance/sdk/pull/681) [`93e7e9d`](https://github.com/enzymefinance/sdk/commit/93e7e9de9d3b4b5e7060c0b45efa83fdc958f821) Thanks [@KedziaPawel](https://github.com/KedziaPawel)! - Add enzyme vault integration
+
 ## 1.2.5
 
 ### Patch Changes