+++ packages/environment/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @enzymefinance/environment
 
+## 6.3.41
+
+### Patch Changes
+
+- [#709](https://github.com/enzymefinance/sdk/pull/709) [`08b8e14`](https://github.com/enzymefinance/sdk/commit/08b8e14f6fb1191b43babc4484ac31346b2884ff) Thanks [@iherger](https://github.com/iherger)! - Morpho
+
+## 6.3.40
+
+### Patch Changes
+
+- [#681](https://github.com/enzymefinance/sdk/pull/681) [`93e7e9d`](https://github.com/enzymefinance/sdk/commit/93e7e9de9d3b4b5e7060c0b45efa83fdc958f821) Thanks [@KedziaPawel](https://github.com/KedziaPawel)! - Add enzyme vault integration
+
+- Updated dependencies [[`93e7e9d`](https://github.com/enzymefinance/sdk/commit/93e7e9de9d3b4b5e7060c0b45efa83fdc958f821)]:
+  - @enzymefinance/sdk@1.2.14
+
 ## 6.3.39
 
 ### Patch Changes